requirements.txt
Flask==3.0.0
Flask-Mail==0.9.1
Flask-Session==0.5.0
python-dotenv==1.0.0
Pillow==10.1.0
pymongo==4.6.1
